hw3

+++ hw3/homework/train_fcn.py
@@ -23,7 +23,7 @@ def accuracy(outputs, labels):
 
 class ClassificationLoss(torch.nn.Module):
     def forward(self, input, target):
-        return F.cross_entropy(input, target)
+        return F.cross_entropy(torch.LongTensor(input), torch.LongTensor(target))
 
 
 def train(args):
